+++ destinations/faros-destination/test/converters/bitbucket.test.ts
@@ -4,10 +4,9 @@ import _ from 'lodash';
 import {getLocal} from 'mockttp';
 import pino from 'pino';
 
-import {tempConfig} from '../temp';
+import {initMockttp, tempConfig} from '../temp';
 import {CLI, read} from './../cli';
 import {bitbucketAllStreamsLog} from './data';
-import {initMockttp} from './mockttp';
 
 describe('bitbucket', () => {
   const logger = pino({
