+++ destinations/faros-destination/test/converters/gitlab.test.ts
@@ -6,7 +6,7 @@ import os from 'os';
 import pino from 'pino';
 
 import {InvalidRecordStrategy} from '../../src';
-import {tempConfig} from '../temp';
+import {initMockttp, tempConfig} from '../temp';
 import {CLI, read} from './../cli';
 import {
   gitlabAllStreamsLog,
@@ -14,7 +14,6 @@ import {
   gitlabPGRawLog,
   readTestResourceFile,
 } from './data';
-import {initMockttp} from './mockttp';
 
 describe('gitlab', () => {
   const logger = pino({
@@ -42,13 +41,13 @@ describe('gitlab', () => {
 
   test('process and write records', async () => {
     await mockttp
-      .post('/graphs/test-graph/models')
+      .forPost('/graphs/test-graph/models')
       .withQuery({schema: 'canonical'})
       .once()
       .thenReply(200, JSON.stringify({}));
 
     await mockttp
-      .post('/graphs/test-graph/revisions')
+      .forPost('/graphs/test-graph/revisions')
       .once()
       .thenReply(
         200,
@@ -60,14 +59,14 @@ describe('gitlab', () => {
 
     let entriesSize = 0;
     await mockttp
-      .post(`/graphs/test-graph/revisions/${revisionId}/entries`)
+      .forPost(`/graphs/test-graph/revisions/${revisionId}/entries`)
       .thenCallback(async (r) => {
         entriesSize = r.body.buffer.length;
         return {statusCode: 204};
       });
 
     await mockttp
-      .patch(`/graphs/test-graph/revisions/${revisionId}`)
+      .forPatch(`/graphs/test-graph/revisions/${revisionId}`)
       .withJsonBodyIncluding({status: 'active'})
       .once()
       .thenReply(204);
