+++ destinations/faros-destination/test/converters/googlecalendar.test.ts
@@ -4,10 +4,9 @@ import _ from 'lodash';
 import {getLocal} from 'mockttp';
 import pino from 'pino';
 
-import {tempConfig} from '../temp';
+import {initMockttp, tempConfig} from '../temp';
 import {CLI, read} from './../cli';
 import {googlecalendarAllStreamsLog} from './data';
-import {initMockttp} from './mockttp';
 
 describe('googlecalendar', () => {
   const logger = pino({
