+++ destinations/faros-destination/test/converters/jenkins.test.ts
@@ -7,9 +7,8 @@ import pino from 'pino';
 
 import {InvalidRecordStrategy} from '../../src';
 import {CLI, read} from '../cli';
-import {tempConfig} from '../temp';
+import {initMockttp, tempConfig} from '../temp';
 import {jenkinsAllStreamsLog, jenkinsLog, readTestResourceFile} from './data';
-import {initMockttp} from './mockttp';
 
 describe('jenkins', () => {
   const logger = pino({
@@ -36,13 +35,13 @@ describe('jenkins', () => {
 
   test('process and write records', async () => {
     await mockttp
-      .post('/graphs/test-graph/models')
+      .forPost('/graphs/test-graph/models')
       .withQuery({schema: 'canonical'})
       .once()
       .thenReply(200, JSON.stringify({}));
 
     await mockttp
-      .post('/graphs/test-graph/revisions')
+      .forPost('/graphs/test-graph/revisions')
       .once()
       .thenReply(
         200,
@@ -54,14 +53,14 @@ describe('jenkins', () => {
 
     let entriesSize = 0;
     await mockttp
-      .post(`/graphs/test-graph/revisions/${revisionId}/entries`)
+      .forPost(`/graphs/test-graph/revisions/${revisionId}/entries`)
       .thenCallback(async (r) => {
         entriesSize = r.body.buffer.length;
         return {statusCode: 204};
       });
 
     await mockttp
-      .patch(`/graphs/test-graph/revisions/${revisionId}`)
+      .forPatch(`/graphs/test-graph/revisions/${revisionId}`)
       .withJsonBodyIncluding({status: 'active'})
       .once()
       .thenReply(204);
